Writing PID file whenever Flamenco-Worker is running.
Added optional creation and removal of a PID file. This only happens when
the `pid` configuration option is given, pointing to the location of the
PID file.

This is an attempt to prevent the sysv script from starting Flamenco-Worker
when it is already running (i.e. after two "service flamenco-worker start"
commands).