Added hack to avoid clogging the logs
Cycles and other render engines produce a line of log for each object in
the scene. In the Spring project there are so many objects (leaves, twigs,
pebbles) that it chokes the Manager with logs.

For now we have some custom code to swallow those lines, in lieu of a
logging system that can handle those volumes properly.