Explicitly return task to the Manager queue when stopping the task
This should solve issues when tasks fail after the worker shuts down or
goes to sleep. Previously the task would be set to 'claimed-by-manager'
when the Worker when to sleep, but flushing the task queue on the Worker
could still trigger reactivation of the task on the Manager. Since the
Worker wouldn't be running the task any more, it would time out. This is
now fixed.