Workaround for FFmpeg not supporting '-pattern_type glob' on Windows
Globbing is the only way in which we can convert arbitrary frame sequences
to a video; the other input options all assume that the frame numbers are
sequential, and stop at the first gap. Globbing just skips gaps and
actually uses all available frames.

The workaround consist of doing the globbing in Python and creating an
index file that lists all the input files.