Added `video-encoding` to the default settings
Since it requires FFmpeg to be available to the Worker, it was disabled
by default. However, it's the only task type that was disabled by default,
and having everything on is probably easier for new installations.
Furthermore, the `blender-render` task type also requires external software
(namely Blender), but that wasn't disabled by default.